:sparkles: Déplacement de composants
