+++ .eslintrc.yaml
@@ -10,7 +10,7 @@ rules:
   react/prop-types: 0
   react/jsx-no-target-blank: 0
   react/no-unescaped-entities: 0
-  react/display-name: 1
+  react/display-name: 0
   react-hooks/rules-of-hooks: error
   react-hooks/exhaustive-deps: warn
 parser: babel-eslint
