+++ source/sites/publicodes/App.js
@@ -7,7 +7,7 @@ import Provider from '../../Provider'
 import RulesList from '../mon-entreprise.fr/pages/Documentation/RulesList'
 import About from './About'
 import Contribution from './Contribution'
-import Day from './Day'
+import Day from './journée/Day'
 import Landing from './Landing'
 import Scenarios from './Scenarios'
 import Simulateur from './Simulateur'
